extract filter to separate KamonFilter for use in Play with compile-time DI
